+++ resources/[dg]/ts-shared/typescript/src/server/classes/jobs.ts
@@ -52,6 +52,9 @@ class Jobs {
   isUserWhitelisted(steamId: string, job: string): Promise<boolean> {
     return global.exports['dg-jobs'].isSteamIdWhitelisted(steamId, job);
   }
+  isCidWhitelisted(cid: number, job: string): boolean {
+    return global.exports['dg-jobs'].isCidWhitelisted(cid, job);
+  }
 
   registerJob(name: string, jobInfo: Jobs.Job) {
     global.exports['dg-jobs'].registerJob(name, jobInfo);