+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Farming items gaan niet meer onder nul quality
 - Restaurant medewerker gaat nu automatisch uit dienst als deze de server verlaat
 - Extra controle op samen materiaal uit smeltoven nemen
+- Check garage access op whitelist en niet op active job
 
 ## [1.0.1]